# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fatourti Mobile Application APIS",
        "description": "API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get all clients for the authenticated user",
                "description": "Retrieves all clients associated with the authenticated user's organization.",
                "operationId": "getClientsForUser",
                "parameters": [
                    {
                        "name": "recherche",
                        "in": "query",
                        "description": "Search query to filter clients by name or identification",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of clients"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create a new client",
                "description": "Creates a new client record in the database.",
                "operationId": "storeClient",
                "requestBody": {
                    "description": "Client details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "identification",
                                    "email",
                                    "phone",
                                    "adresse"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Salim Yaa"
                                    },
                                    "identification": {
                                        "type": "string",
                                        "example": "ID123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "yaa@exa.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "123 Main St, City, Country"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get a specific client by ID",
                "description": "Retrieves a specific client's details.",
                "operationId": "showClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client details"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update a specific client by ID",
                "description": "Updates an existing client record in the database.",
                "operationId": "updateClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Client details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Salim Yaa"
                                    },
                                    "identification": {
                                        "type": "string",
                                        "example": "ID123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "yaa@exa.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "123 Main St, City, Country"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client updated successfully"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete a specific client by ID",
                "description": "Deletes a specific client record from the database.",
                "operationId": "deleteClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Client not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}